Client Contract Form

+++ IBS/Models/ClientContractModel.cs
@@ -15,7 +15,7 @@ namespace IBS.Models
 
         public string Client { get; set; } = null!;
 
-        public byte RitesOfficerCd { get; set; }
+        public int RitesOfficerCd { get; set; }
 
         public string? Highlights { get; set; }
